move clean up to Session cleanupSession
