+++ src/libs/actions/Session/index.ts
@@ -4,14 +4,20 @@ import {ChannelAuthorizationCallback} from 'pusher-js/with-encryption';
 import {Linking} from 'react-native';
 import Onyx, {OnyxUpdate} from 'react-native-onyx';
 import {ValueOf} from 'type-fest';
+import * as PersistedRequests from '@libs/actions/PersistedRequests';
 import * as API from '@libs/API';
 import * as Authentication from '@libs/Authentication';
 import * as ErrorUtils from '@libs/ErrorUtils';
+import HttpUtils from '@libs/HttpUtils';
 import Log from '@libs/Log';
 import Navigation from '@libs/Navigation/Navigation';
+import navigationRef from '@libs/Navigation/navigationRef';
+import * as MainQueue from '@libs/Network/MainQueue';
 import * as NetworkStore from '@libs/Network/NetworkStore';
+import NetworkConnection from '@libs/NetworkConnection';
 import * as Pusher from '@libs/Pusher/pusher';
 import * as ReportUtils from '@libs/ReportUtils';
+import * as SessionUtils from '@libs/SessionUtils';
 import Timers from '@libs/Timers';
 import {hideContextMenu} from '@pages/home/report/ContextMenu/ReportActionContextMenu';
 import * as Device from '@userActions/Device';
@@ -23,6 +29,7 @@ import CONFIG from '@src/CONFIG';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
+import SCREENS from '@src/SCREENS';
 import Credentials from '@src/types/onyx/Credentials';
 import {AutoAuthState} from '@src/types/onyx/Session';
 import clearCache from './clearCache';
@@ -584,13 +591,39 @@ function clearSignInData() {
 }
 
 /**
+ * Reset all current params of the Home route
+ */
+function resetHomeRouteParams() {
+    Navigation.isNavigationReady().then(() => {
+        const routes = navigationRef.current?.getState().routes;
+        const homeRoute = routes?.find((route) => route.name === SCREENS.HOME);
+
+        const emptyParams: Record<string, undefined> = {};
+        Object.keys(homeRoute?.params ?? {}).forEach((paramKey) => {
+            emptyParams[paramKey] = undefined;
+        });
+
+        Navigation.setParams(emptyParams, homeRoute?.key ?? '');
+        Onyx.set(ONYXKEYS.IS_CHECKING_PUBLIC_ROOM, false);
+    });
+}
+
+/**
  * Put any logic that needs to run when we are signed out here. This can be triggered when the current tab or another tab signs out.
+ * - Cancels pending network calls - any lingering requests are discarded to prevent unwanted storage writes
+ * - Clears all current params of the Home route - the login page URL should not contain any parameter
  */
 function cleanupSession() {
     Pusher.disconnect();
     Timers.clearAll();
     Welcome.resetReadyCheck();
     PriorityMode.resetHasReadRequiredDataFromStorage();
+    MainQueue.clear();
+    HttpUtils.cancelPendingRequests();
+    PersistedRequests.clear();
+    NetworkConnection.clearReconnectionCallbacks();
+    SessionUtils.resetDidUserLogInDuringSession();
+    resetHomeRouteParams();
 }
 
 function clearAccountMessages() {
